requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Django==2.2.3
djangorestframework==3.9.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
Pillow==6.1.0
pycodestyle==2.5.0
Pygments==2.4.2
pylint==2.3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
